pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.10.0",
    "langchain-community>=0.3.18",
    "langchain>=0.3.19",
    "openai>=1.64.0",
    "plotly>=6.0.0",
    "streamlit>=1.42.2",
]
